# Conflicts:
#	src/antares/craft/exceptions/exceptions.py
#	src/antares/craft/service/api_services/models/st_storage.py
#	src/antares/craft/service/local_services/factory.py

+++ src/antares/craft/model/study.py
@@ -26,6 +26,7 @@ from antares.craft import (
 from antares.craft.exceptions.exceptions import (
     LinkCreationError,
     ReadingMethodUsedOufOfScopeError,
+    ReferencedObjectDeletionNotAllowed,
     UnsupportedStudyVersion,
 )
 from antares.craft.model.area import Area, AreaProperties, AreaPropertiesUpdate, AreaUi
@@ -33,7 +34,9 @@ from antares.craft.model.binding_constraint import (
     BindingConstraint,
     BindingConstraintProperties,
     BindingConstraintPropertiesUpdate,
+    ClusterData,
     ConstraintTerm,
+    LinkData,
 )
 from antares.craft.model.link import Link, LinkProperties, LinkPropertiesUpdate, LinkUi
 from antares.craft.model.output import Output
@@ -77,10 +80,10 @@ class Study:
         self._binding_constraints_service = services.bc_service
         self._settings_service = services.settings_service
         self._settings = StudySettings()
-        self._areas: dict[str, Area] = dict()
-        self._links: dict[str, Link] = dict()
-        self._binding_constraints: dict[str, BindingConstraint] = dict()
-        self._outputs: dict[str, Output] = dict()
+        self._areas: dict[str, Area] = {}
+        self._links: dict[str, Link] = {}
+        self._binding_constraints: dict[str, BindingConstraint] = {}
+        self._outputs: dict[str, Output] = {}
         self._solver_path: Optional[Path] = solver_path
 
         study_version = StudyVersion.parse(version)
@@ -146,6 +149,20 @@ class Study:
         return area
 
     def delete_area(self, area: Area) -> None:
+        # Check area is not referenced in any binding constraint
+        referencing_binding_constraints = []
+        for bc in self._binding_constraints.values():
+            for term in bc._terms.values():
+                data = term.data
+                if (isinstance(data, ClusterData) and data.area == area.id) or (
+                    isinstance(data, LinkData) and (data.area1 == area.id or data.area2 == area.id)
+                ):
+                    referencing_binding_constraints.append(bc.name)
+                    break
+        if referencing_binding_constraints:
+            raise ReferencedObjectDeletionNotAllowed(area.id, referencing_binding_constraints, object_type="Area")
+
+        # Delete the area
         self._area_service.delete_area(area.id)
         self._areas.pop(area.id)
 
@@ -177,6 +194,18 @@ class Study:
         return link
 
     def delete_link(self, link: Link) -> None:
+        # Check link is not referenced in any binding constraint
+        referencing_binding_constraints = []
+        for bc in self._binding_constraints.values():
+            for term in bc._terms.values():
+                data = term.data
+                if isinstance(data, LinkData) and data.area1 == link.area_from_id and data.area2 == link.area_to_id:
+                    referencing_binding_constraints.append(bc.name)
+                    break
+        if referencing_binding_constraints:
+            raise ReferencedObjectDeletionNotAllowed(link.id, referencing_binding_constraints, object_type="Link")
+
+        # Delete the link
         self._link_service.delete_link(link)
         self._links.pop(link.id)
